6 Patterns completed.
1) Strategy design pattern completed.
Important comments added to the code in Turkish.

2) Observer design pattern completed.
Important comments added to the code in Turkish

3)  Chain of responsibility design pattern completed.
Important comments added to the code in Turkish

3)  Memento design pattern completed.
Important comments added to the code in Turkish

4)   Template Method design pattern completed.
Important comments added to the code in Turkish

5) State design pattern completed.
Important comments added to the code in Turkish

6) Mediator design pattern completed.
Important comments added to the code in Turkish

Finally, project for the Command Design Pattern created. Not implemented.